Move shared classes from human to controls package
