+++ TelepresenceBot/mobile/src/main/java/com/novoda/tpbot/controls/CommandRepeater.java
@@ -1,4 +1,4 @@
-package com.novoda.tpbot.human;
+package com.novoda.tpbot.controls;
 
 import android.os.Handler;
