+++ TelepresenceBot/mobile/src/main/res/layout/switchable_controller.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.novoda.tpbot.human.ControllerView xmlns:android="http://schemas.android.com/apk/res/android"
+<com.novoda.tpbot.controls.ControllerView xmlns:android="http://schemas.android.com/apk/res/android"
   android:id="@+id/bot_controller_direction_view"
   android:layout_width="match_parent"
   android:layout_height="match_parent"
